Fix (long) cast on file handle -- HOUSE RENDERS!
fr_read (and every other GEMDOS Fread/Fclose/Fwrite call) was
casting the short file handle to (long):

    err = _gemdos(GEMDOS_Fread, (long) fileHandle, count, (long) buffer);

Alcyon's variadic _gemdos wrapper pushes each argument at its natural
type width, so the (long) cast pushed 4 bytes where TOS expects 2.
The trap handler then read:
    SP+0 : func code (short) = 0x3F -- OK
    SP+2 : handle (short) = HIGH word of the long handle = 0x0000
    SP+4 : count (long) = (fh_lo << 16) | (count_hi) -- corrupted
    SP+8 : buffer (long) = (count_lo << 16) | ... -- corrupted

So TOS was silently reading from handle 0 (stdin) with a huge bogus
count, returning 0 bytes for every call.  Files were "opening"
successfully but Fread was never actually reading anything -- the
buffers stayed at whatever pattern they had after BSS init and
GEMDOS_Malloc filling.

Explanation for the striped house: our LCP sprite draws looked
plausible because sp_lbal generated deterministic pixel data
from the (mostly-zero) body_lcp_buffer.  Same for the house
compositor -- it faithfully blk-copied a uniform pattern from
the untouched g_srptr buffer.

Fix: pass fileHandle as its natural short type across every
_gemdos(GEMDOS_Fread/Fclose/Fwrite/...) call.  Bulk-patched save.c,
assets.c, letload.c, cards.c, sound.c.

Result: HOUSE.SCN decompresses correctly, blkcopy32 copies real
house pixels into the compositor, and the ST shifter shows the
full LCP living room -- yellow-walled kitchen, blue-walled
bedroom, cyan bathroom, TV / stereo / plants / staircase, dining
table -- matching the reference screenshot from a real ST run.

Palette is slightly off (yellow walls instead of green) but the
scene structure is exact.  That's the next thing to look at once
we chase the remaining check_for_any_action_triggers NULL deref
and the palette applies_clothing_colors init function.

+++ csrc/assets.c
@@ -77,7 +77,7 @@ load_objects()
 
         fileHandle = file_open("objects", 0);
         fr_read(fileHandle, 14000L, objects_file);
-        _gemdos(GEMDOS_Fclose, (long) fileHandle, 0L, 0L);
+        _gemdos(GEMDOS_Fclose, fileHandle, 0L, 0L);
 }
 
 /* load_sprites: read the 14000-byte SPRITES file into sprites_files[].
@@ -90,7 +90,7 @@ load_sprites()
 
         fileHandle = file_open("sprites", 0);
         fr_read(fileHandle, 14000L, sprites_files);
-        _gemdos(GEMDOS_Fclose, (long) fileHandle, 0L, 0L);
+        _gemdos(GEMDOS_Fclose, fileHandle, 0L, 0L);
 }
 
 /* Parse a sequence-of-records buffer (OBJECTS or SPRITES format) into
@@ -188,7 +188,7 @@ long            max_bytes;
         if (total > max_bytes)
                 total = max_bytes;
         fr_read(fileHandle, total, dest_buf);
-        _gemdos(GEMDOS_Fclose, (long) fileHandle, 0L, 0L);
+        _gemdos(GEMDOS_Fclose, fileHandle, 0L, 0L);
         return count;
 }
 
@@ -421,7 +421,7 @@ long            dest_size_words;
                 _gemdos(GEMDOS_Mfree, (long) nbuf, 0L, 0L);
         }
 
-        _gemdos(GEMDOS_Fclose, (long) filehandle, 0L, 0L);
+        _gemdos(GEMDOS_Fclose, filehandle, 0L, 0L);
         _gemdos(GEMDOS_Mfree,  (long) fbuffer_orig, 0L, 0L);
 }
 
@@ -440,6 +440,6 @@ long            max_bytes;
 
         fileHandle = file_open("names", 0);
         fr_read(fileHandle, max_bytes, dest_buf);
-        _gemdos(GEMDOS_Fclose, (long) fileHandle, 0L, 0L);
+        _gemdos(GEMDOS_Fclose, fileHandle, 0L, 0L);
         return max_bytes;
 }
